digital-netra-backend/pyproject.toml
[tool.poetry]
name = "digital-netra-backend"
version = "0.1.0"
description = "Digital Netra authentication backend"
authors = ["Digital Netra"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.135.0"
uvicorn = "^0.23.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.6"
alembic = "^1.12.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.0"
bcrypt = "^4.1.0"
PyJWT = "^2.8.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
